Require graphiql-rails gem in development
This needs to be required here because of the peculiarities of how
engines hook into the rails application. Essentially, requiring an
engine within an engine will cause this to fail - although the library
is loaded correctly, the app code never gets loaded causing a 404
error when performing a GET to /graphql/explorer.