.to_s ALL THE PATHNAMES IN TEH $LOAD_PATH !!1!
As part of fixing the memory present in ruby 2.3 and 2.4:

    https://bugs.ruby-lang.org/issues/14372

There were specs added to ensure that the memory leak was not
re-introduced into the server:

    spec/lib/ruby_require_leak_spec.rb

Without these changes, this spec would randomly fail if one of these
tool specs was run prior to the spec/lib/ruby_require_leak_spec.rb file
when running the entire test suite.

This change makes it so their isn't a false positive test failure for
the ruby_require_leak_spec.rb.  This doesn't need to be backported,
though, since the tools are never used in the server code.