Exclude Service::AGGREGATE_ALL_VM_ATTRS from MiqExp.to_sql
These calls, while they are Arel and could be converted, would most
likely be used in "filter" type scenarios, and would not be good
candidates for that since they would perform horribly when used on more
than one record at a time (which filtering would do that).

Instead of preventing these records from showing up as a filter option,
just make sure they can't be used with SQL when they are.  That said,
they will cause N+1 hell, so they really shouldn't be filtered on...
but...

¯\(°_o)/¯