Add miqssh utilities to tools dir.
These tools are very helpful for managing multiple appliances.
The miqgrep and miqtail commands are useful for single appliances as well when used with -r option to follow particular requests and associated tasks in the MIQ logs.
Additional details on the utilities are available in the README.md file.
See INSTALL.txt for setting this tool up for use the first time.