Lookup a category_id if a tag control passes it in
When editing a tag control for service dialogs, we were always looking up by the name
coming in from the dialog.  For a new record this sufficed, for an edit it caused edits to fail.

This changes the order of lookups to first check the category_id if it comes in,
and to fallback to the name if there is no category_id

https://bugzilla.redhat.com/show_bug.cgi?id=1542590