Allow OrchestraionTemplate subclass to customize md5 calculation
Currently, md5 checksum is forcibly calculated on entire orchestration
template text content:

```
OrchestrationTemplate.calc_md5(text)
```

But this is not always feasible. VMware vCloud, for example, yields
randomly ordered elements in XML even when describing the very same orchestration
template, hence the MD5 checksum differs. Therefore we need to be able to
customize the way the MD5 checksum is calculated.

With this commit we make sure that calc_md5 function from the subclass
is used:

```
ManageIQ::Providers::Vmware::CloudManager::OrchestrationTemplate.calc_md5(text)
```

Only this way, we're able to customize what part of the XML will actually
be used to calculate MD5 checksum.

Signed-off-by: Miha Pleško <miha.plesko@xlab.si>