Use ActiveSupport::Logger.broadcast instead of MulticastLogger
Makes a switch to use `ActiveSupport::Logger.broadcast` which is already
used to by the `rails console` to handle printing to both STDOUT and the
terminal for the rails log.  This just makes use of the same
functionality that is there that is pretty much being replicated in the
MulticastLogger that we put together ourselves.

Minor difference that shouldn't matter:  We now only control the level
of the main logger, but that really didn't matter since we don't
multicast to more than two loggers, and the top one is the one is the
only one that matters.  The `$container_log` will always be debug
anyway, so no need for the level wrap that we had in `MulticastLogger`.