Set job data retention at runtime in EmbeddedAnsibleWorker::Runner#do_work
We don't really want to do this in the Config::Activator because
that will cause every worker to run through the same code as they
sync their config.

We keep a local cache (read: instance variable) of the last value
we set in tower so that we're not constantly hitting the API to
determine if the value changed.

The EmbeddedAnsibleWorker can use ::Settings directly because it
is threaded and the server process takes care of reloading it.

https://bugzilla.redhat.com/show_bug.cgi?id=1560478