Move get_file and save_file to ConfigurationManagementMixin
In moving, the methods become settings_for_resource_yaml and
add_settings_for_resource_yaml, respectively.  The old interfaces are
still left intact temporarily until the UI is changed to use the new
methods.

Also preserved is the special activation that is currently done for
ntp_reload, as introduced in ManageIQ/manageiq@939524fa, which avoids
repeated activation for every worker, instead preferring to do it once
at the server level.

https://bugzilla.redhat.com/show_bug.cgi?id=1082155